Fixed build on linux64 - it was really bad that behavior of macro FB_ALIGN depended upon alignment parameter type
git-svn-id: svn+ssh://svn.code.sf.net/p/firebird/code/firebird/trunk@59899 65644016-39b1-43b1-bf79-96bc8fe82c15

+++ src/include/types_pub.h
@@ -45,7 +45,7 @@ typedef unsigned long uintptr_t;
 
 #endif
 
-#define FB_ALIGN(n, b) ((n + b - 1) & ~(b - 1))
+#define FB_ALIGN(n, b) ((n + size_t(b - 1)) & ~size_t(b - 1))
 
 /******************************************************************/
 /* API handles                                                    */
